testbed/docker/api/requirements.txt
fastapi
uvicorn[standard]
strawberry-graphql[fastapi]
psycopg2-binary
pydantic<2.12
